services/deploy-chroma/pyproject.toml
[tool.poetry]
name = "deploy-chroma"
version = "0.1.0"
description = ""
authors = ["Kyle Lee <kyle@vessl.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
chromadb = "^0.5.5"
bcrypt = "^4.2.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
